second commit

+++ src/main/java/com/ylc/hhtally/pojo/Bill.java
@@ -11,7 +11,7 @@ public class Bill {
     private int billId;
     private int userId;
     private int labelId;
-    private int money;
+    private double money;
     private String time;
     private String remark;
     private String shopkeeper;
